# swagger.yml
openapi: 3.0.3
info:
  title: Msgtrik Chat Application API
  description: |
    This is the API documentation for the Msgtrik Chat Application.
    It covers all endpoints needed for authentication, user management, and messaging.
  version: 1.0.0
  contact:
    email: support@msgtrik.example.com
servers:
  - url: https://api.msgtrik.example.com/v1
    description: Production server
  - url: https://staging-api.msgtrik.example.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          example: 400
        message:
          type: string
          example: "Bad request"
        details:
          type: string
          example: "Email already exists"
paths:
...
  /api/users/token/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh JWT token
      description: Obtain a new access token using a valid refresh token.
      operationId: refreshToken
      requestBody:
        description: Refresh token
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh:
                  type: string
                  description: The refresh token issued during login
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: New access token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  access:
                    type: string
                    description: The new access token
                    example: "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...